requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
colorlog==6.9.0
cryptography==46.0.2
fastapi==0.115.11
h11==0.16.0
idna==3.10
pip==24.0
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.46.2
sqlalchemy==2.0.44
typing_extensions==4.12.2
uvicorn==0.24.0.post1
dnspython==2.8.0
email-validator==2.3.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.2
httpx==0.25.2
faker==19.12.0
reportlab==4.4.5